Update styles for row and squares
Row was not flexing correctly due to container having
both justify-content and align-items to be center.

To fix, these have been removed and changed flex direction to
have rows be in actual rows instead of columns. Squares have
been changed to white and take up the whole space of the container.